NP-1997 Move buttons for rejecting/creating DOI
