+++ src/pages/public_registration/PublicRegistrationStatusBar.tsx
@@ -1,10 +1,12 @@
 import React, { FC, useState } from 'react';
 import { useSelector, useDispatch } from 'react-redux';
 import styled from 'styled-components';
-import { Button, DialogActions, TextField, Tooltip, Typography } from '@material-ui/core';
+import { Button, DialogActions, TextField, Typography } from '@material-ui/core';
 import CheckCircleOutlineIcon from '@material-ui/icons/CheckCircleOutline';
 import WorkOutlineIcon from '@material-ui/icons/WorkOutline';
 import CloudUploadIcon from '@material-ui/icons/CloudUpload';
+import CloseIcon from '@material-ui/icons/Close';
+import CheckIcon from '@material-ui/icons/Check';
 import { useTranslation } from 'react-i18next';
 import { Link } from 'react-router-dom';
 
@@ -16,7 +18,7 @@ import { setNotification } from '../../redux/actions/notificationActions';
 import { NotificationVariant } from '../../types/notification.types';
 import ButtonWithProgress from '../../components/ButtonWithProgress';
 import { RegistrationStatus, DoiRequestStatus } from '../../types/registration.types';
-import { createDoiRequest } from '../../api/doiRequestApi';
+import { createDoiRequest, updateDoiRequest } from '../../api/doiRequestApi';
 import { publishRegistration } from '../../api/registrationApi';
 import { registrationValidationSchema } from '../../utils/validation/registration/registrationValidation';
 
@@ -49,6 +51,8 @@ enum LoadingName {
   None = '',
   Doi = 'DOI',
   Publish = 'PUBLISH',
+  RejectDoi = 'REJECT_DOI',
+  ApproveDoi = 'APPROVE_DOI',
 }
 
 export const PublicRegistrationStatusBar: FC<PublicRegistrationContentProps> = ({
@@ -86,7 +90,23 @@ export const PublicRegistrationStatusBar: FC<PublicRegistrationContentProps> = (
         dispatch(setNotification(t('feedback:success.doi_request_sent')));
       }
     }
-    setIsLoading(LoadingName.None);
+  };
+
+  const onClickUpdateDoiRequest = async (status: DoiRequestStatus) => {
+    if (status === DoiRequestStatus.Approved) {
+      setIsLoading(LoadingName.ApproveDoi);
+    } else {
+      setIsLoading(LoadingName.RejectDoi);
+    }
+    const updateDoiResponse = await updateDoiRequest(identifier, status);
+    if (updateDoiResponse) {
+      if (updateDoiResponse.error) {
+        dispatch(setNotification(t('feedback:error.update_doi_request'), NotificationVariant.Error));
+      } else {
+        refetchRegistration();
+        dispatch(setNotification(t('feedback:success.doi_request_updated'), NotificationVariant.Success));
+      }
+    }
   };
 
   const onClickPublish = async () => {
@@ -100,11 +120,10 @@ export const PublicRegistrationStatusBar: FC<PublicRegistrationContentProps> = (
         dispatch(setNotification(t('feedback:success.published_registration'), NotificationVariant.Success));
       }
     }
-    setIsLoading(LoadingName.None);
   };
 
-  const isOwner = user.isCreator && owner === user.id;
-  const isCurator = user.isCurator && user.customerId === publisher.id;
+  const isOwner = user && user.isCreator && owner === user.id;
+  const isCurator = user && user.isCurator && user.customerId === publisher.id;
   const hasNvaDoi = !!doi;
 
   return isOwner || isCurator ? (
@@ -131,9 +150,34 @@ export const PublicRegistrationStatusBar: FC<PublicRegistrationContentProps> = (
 
         {!hasNvaDoi &&
           (doiRequest?.status === DoiRequestStatus.Requested ? (
-            <Button variant="contained" color="primary" disabled>
-              {t('public_page.requested_doi')}
-            </Button>
+            user.isCurator ? (
+              <>
+                <ButtonWithProgress
+                  color="primary"
+                  variant="contained"
+                  data-testid="button-reject-doi"
+                  endIcon={<CloseIcon />}
+                  onClick={() => onClickUpdateDoiRequest(DoiRequestStatus.Rejected)}
+                  isLoading={isLoading === LoadingName.RejectDoi}
+                  disabled={!!isLoading}>
+                  {t('common:reject_doi')}
+                </ButtonWithProgress>
+                <ButtonWithProgress
+                  color="primary"
+                  variant="contained"
+                  data-testid="button-create-doi"
+                  endIcon={<CheckIcon />}
+                  onClick={() => onClickUpdateDoiRequest(DoiRequestStatus.Approved)}
+                  isLoading={isLoading === LoadingName.ApproveDoi}
+                  disabled={!!isLoading || !registrationIsValid}>
+                  {t('common:create_doi')}
+                </ButtonWithProgress>
+              </>
+            ) : (
+              <Button variant="contained" color="primary" disabled>
+                {t('public_page.requested_doi')}
+              </Button>
+            )
           ) : (
             <>
               {status === RegistrationStatus.PUBLISHED && (
@@ -154,18 +198,18 @@ export const PublicRegistrationStatusBar: FC<PublicRegistrationContentProps> = (
           ))}
 
         {status === RegistrationStatus.DRAFT && (
-          <Tooltip arrow title={<Typography>{t('public_page.fix_validation_errors_before_publishing')}</Typography>}>
-            <span>
-              <ButtonWithProgress
-                disabled={!!isLoading || !registrationIsValid}
-                data-testid="button-publish-registration"
-                endIcon={<CloudUploadIcon />}
-                onClick={onClickPublish}
-                isLoading={isLoading === LoadingName.Publish}>
-                {t('common:publish')}
-              </ButtonWithProgress>
-            </span>
-          </Tooltip>
+          // <Tooltip arrow title={<Typography>{t('public_page.fix_validation_errors_before_publishing')}</Typography>}>
+          <span>
+            <ButtonWithProgress
+              disabled={!!isLoading || !registrationIsValid}
+              data-testid="button-publish-registration"
+              endIcon={<CloudUploadIcon />}
+              onClick={onClickPublish}
+              isLoading={isLoading === LoadingName.Publish}>
+              {t('common:publish')}
+            </ButtonWithProgress>
+          </span>
+          // </Tooltip>
         )}
       </div>
